lib: Add ucontext_t for all architectures.
ucontext_t is used by `tcc -run` and it affects the compilation process
of that.  If we don't have the data structures defined the build process
of tcc-mob later in the chain fails.  We don't use the `-run` option for
anything in the bootstrapping process, but as we do `ONE_SOURCE` that
affect our building process.  The cleanest solution is to add the proper
support here and forget about it.

This commit copies the code from musl libc, using the GNU_SOURCE part of
the `#ifdef`s in there as that's what tcc expects.  We can do better but
the goal of the commit is to make the bare minimum.

* include/signal.h (ucontext_t): Move current i386 support to...
* include/gnu/x86/signal.h: ...this new file.
* include/linux/x86/signal.h,
include/linux/x86_64/signal.h,
include/linux/arm/signal.h,
include/linux/riscv64/signal.h: New files.
* simple.sh: Add support for multiple signal.h files.
* configure.sh: Likewise.
* configure: Likewise.