mescc: Fix length calculation of initialized array with unspecified length.
Consider the following definition:

    char a[] = {1, 2, 3};

The initializer is parsed as

    ((initzer (initzer-list
        (initzer (p-expr (fixed "1")))
        (initzer (p-expr (fixed "2")))
        (initzer (p-expr (fixed "3"))))))

so (cadar init) is

    (initzer-list
        (initzer (p-expr (fixed "1")))
        (initzer (p-expr (fixed "2")))
        (initzer (p-expr (fixed "3"))))

We want to determine the number of items in the initializer, which
is (length (cdadar init)), not (length (cadar init)).

Using cadar causes the length of the array to be one too large and
incorrect evaluation of sizeof.

* module/mescc/compile.scm (init-declr->info): Fix length calculation
of initialized array with unspecified length.