benchmark: Lay out the skeleton.
This mini-benchmark uses `time` to check the time needed to run certain
actions: a fibonacci sequence, compiling a minimal file with mescc and
compiling mes itself with mescc.

The first run shows guile is around 90x faster than mes in these tasks,
so there's some room to improvement.

We could improve this by running the benchmarks several times and obtain
averages and standard deviations in order to make this more academic,
but taking in account the numbers we are dealing with, that shouldn't be
a concern at the moment.

* benchmarks/fib.scm, benchmarks/mescc-hello.c: Add files.
* guix/git/mes.scm (native-inputs): Add `time`.
* simple.make (benchmark*): Add rules.
(clean): Add benchmark/*.
(TIME, TIME_FMT): New variables.
* .gitignore: Add benchmarks/tmp and benchmarks/out.